+++ sdk/src/test_utils/contracts.rs
@@ -1,6 +1,6 @@
 use {
     crate::sui::{self, traits::*},
-    std::path::PathBuf,
+    std::{fs::OpenOptions, path::PathBuf},
     sui_move_build::implicit_deps,
     sui_package_management::system_package_versions::latest_system_packages,
 };
@@ -91,6 +91,13 @@ pub async fn publish_move_package(
             panic!("Transaction has erroneous effects: {path_str} {effects}");
         }
     }
+
+    // Create the lock file if not exists.
+    OpenOptions::new()
+        .write(true)
+        .create(true)
+        .open(&lock_file)
+        .expect("Failed to create lock file.");
 
     sui_package_management::update_lock_file(
         wallet,