+++ toolkit-rust/src/nexus_tool.rs
@@ -60,7 +60,7 @@ pub trait NexusTool: Send + Sync + 'static {
     /// It is used to generate the `/health` endpoint.
     fn health(&self) -> impl Future<Output = AnyResult<StatusCode>> + Send;
     /// Returns the relative path on a webserver that the tool resides on. This
-    /// defaults to an empty path (root URL). But can be overriden by the
+    /// defaults to an empty path (root URL). But can be overridden by the
     /// implementor.
     fn path() -> &'static str {
         ""