+++ tools/llm-openai-chat-completion/src/main.rs
@@ -120,10 +120,10 @@ enum MessageKind {
     System,
     Assistant,
     Tool,
-    Funtion,
+    Function,
 }
 
-/// Attemps to convert a [`String`] to a [`MessageKind`].
+/// Attempts to convert a [`String`] to a [`MessageKind`].
 impl TryFrom<String> for MessageKind {
     type Error = anyhow::Error;
 
@@ -141,7 +141,7 @@ impl From<Role> for MessageKind {
             Role::System => MessageKind::System,
             Role::User => MessageKind::User,
             Role::Assistant => MessageKind::Assistant,
-            Role::Function => MessageKind::Funtion,
+            Role::Function => MessageKind::Function,
             Role::Tool => MessageKind::Tool,
         }
     }
@@ -553,7 +553,7 @@ mod tests {
 
         let role = Role::Function;
         let kind: MessageKind = role.into();
-        assert_eq!(kind, MessageKind::Funtion);
+        assert_eq!(kind, MessageKind::Function);
 
         let role = Role::Tool;
         let kind: MessageKind = role.into();