+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [`0.2.0`] - Unreleased
+
+### `nexus-cli`
+
+#### Changed
+
+- JSON DAG definition no longer specifies entry input ports
+- renamed JSON DAG `vertices.input_ports` to `vertices.entry_ports`
+
+### `nexus-sdk`
+
+#### Fixed
+
+- `test_utils::contracts` now creates a `Move.lock` if it doesn't exist yet
+
 ## [`0.1.0`] - 2025-04-14
 
 ### `nexus-cli`
@@ -31,10 +46,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 
-- added basic strcuture for Nexus Tools written in Rust in the form of a trait
+- added basic structure for Nexus Tools written in Rust in the form of a trait
 - added a macro that starts a webserver for one or multiple tools, providing all necessary endpoints
 - added a first, dumb version of secret manager
-- added a dev guide that goes through the steps to use CLI to scaffold a boilerplate tool and impelement NexusTool trait
+- added a dev guide that goes through the steps to use CLI to scaffold a boilerplate tool and implement NexusTool trait
 
 ### `nexus-sdk`